Create DISCLAIMER.md
# DISCLAIMER
Use of this service is limited only to **non-sensitive and publicly available
data**. Users must not use, share, or store any kind of sensitive data like
health status, provision or payment of healthcare, Personally Identifiable
Information (PII) and/or Protected Health Information (PHI), etc. under **ANY**
circumstance.

Administrators for this service reserve the right to moderate all information
used, shared, or stored with this service at any time. Any user that cannot
abide by this disclaimer and Code of Conduct  may be subject to action, up to
and including revoking access to services.

The material embodied in this software is provided to you "as-is" and without
warranty of any kind, express, implied or otherwise, including without
limitation, any warranty of fitness for a particular purpose. In no event shall
the Centers for Disease Control and Prevention (CDC) or the United States (U.S.)
government be liable to you or anyone else for any direct, special, incidental,
indirect or consequential damages of any kind, or any damages whatsoever,
including without limitation, loss of profit, loss of use, savings or revenue,
or the claims of third parties, whether or not CDC or the U.S. government has
been advised of the possibility of such loss, however caused and on any theory
of liability, arising out of or in connection with the possession, use or
performance of this software.